% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert between *d*, *r* and *Odds ratio*}
\usage{
convert(
  d = NULL,
  r = NULL,
  r2 = NULL,
  f = NULL,
  f2 = NULL,
  eta2 = NULL,
  ods = NULL,
  log.ods = NULL,
  digits = 2
)
}
\arguments{
\item{d}{Cohen's d}

\item{r}{Pearson's correlation r}

\item{r2}{R-squared}

\item{f, f2}{Cohen's f}

\item{eta2}{eta}

\item{ods}{Odds ratio (OR)}

\item{log.ods}{Log odds ratio}

\item{digits}{digits}
}
\value{
vector
}
\description{
Convert between *d*, *r* and *Odds ratio*
}
\examples{

 convert(d=.30, digits= 4)
convert(r=.14834)
convert(r2=.0220)
convert(f=.1500 )
convert(f2=.0225)
convert(eta2=.0220)
convert(ods=1.7231)
convert(log.ods=0.5441)
}
