% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop-out.R
\name{drop_out_rate}
\alias{drop_out_rate}
\title{Drop-out rate}
\usage{
drop_out_rate(n, d = 0.1)
}
\arguments{
\item{n}{is the sample size required}

\item{d}{is the dropout rate}
}
\value{
vector
}
\description{
Drop-out rate
}
\examples{

drop_out_rate(100, 1/20)
}
