% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin-error.R
\name{margin_error}
\alias{margin_error}
\title{margin of error}
\usage{
margin_error(
  n = NULL,
  margin.error = NULL,
  p = NULL,
  ci = 0.95,
  sd = NULL,
  alternative = c("two.sided", "less", "greater"),
  names = NULL,
  type = c("proportion", "mean"),
  main = "Point Estimate of Population Proportion",
  digits = 3
)
}
\arguments{
\item{n}{Total number of observations}

\item{margin.error}{ME}

\item{p}{probabiliti}

\item{ci}{confidenz Intervall}

\item{sd}{Streuung für Mittelwert}

\item{alternative}{alternative hypothesis}

\item{names}{erste Spalte}

\item{type}{"proportion", "mean"}

\item{main}{text ausgabe}

\item{digits}{digits}
}
\value{
data.frame
}
\description{
margin of error
}
\examples{

# haendische Kontrolle

z<- 1.96
p <- .5
n<- 100
round((me = z*(sqrt(p*(1-p) / n))),3)

round(  p + c(-me, me),3)
margin_error(margin.error =me, p = c(.5, .4) )

margin_error(margin.error = .025, p = .50)
margin_error(n = 1537, p = .50)
margin_error(margin.error = .025, n = 1536.584)

#' margin_error(sd = 100, margin.error = 25)
margin_error(n = 62, margin.error = 25, type="mean")
margin_error(n = 62, sd = 100)


margin_error(sd = c(10, 15), margin.error = 2.5)
margin_error(margin.error = .025, n = 1536.584)
# 

margin_error(margin.error = .025, p = c(.50, .60))
margin_error(margin.error = .025, n = c(1500, 2000))

}
