% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{sample_size}
\alias{sample_size}
\alias{sample_size.character}
\alias{sample_size.power.htest}
\alias{sample_size.webpower}
\title{Sample Size}
\usage{
sample_size(x, ...)

\method{sample_size}{character}(
  x = c("lm", "glm", "cor", "lmer", "aov", "t.test", "chisq.test", "wilcox.test"),
  treatment = NULL,
  control = 1,
  ndf = treatment - 1,
  ng = (ndf + 1) * control,
  measurements = 1,
  nm = measurements,
  drop.out.rate = NULL,
  nscor = 1,
  type = "two.sample",
  predictors = NULL,
  df = predictors,
  u = df,
  v = NULL,
  d = c(small = 0.2, medium = 0.5, large = 0.8),
  r = c(small = 0.1, medium = 0.3, large = 0.5),
  f = c(small = 0.1, medium = 0.25, large = 0.4),
  f2 = c(small = 0.02, medium = 0.15, large = 0.35),
  h = c(small = 0.2, medium = 0.5, large = 0.8),
  w = c(small = 0.1, medium = 0.3, large = 0.5),
  power = 0.8,
  sig.level = 0.05,
  alternative = "two.sided"
)

\method{sample_size}{power.htest}(x)

\method{sample_size}{webpower}(x)
}
\arguments{
\item{x}{Test name or pwr / WebPower object}

\item{...}{}

\item{treatment, ndf, }{Number of treatment-groups}

\item{control}{Number of control-groups}

\item{ng}{treatment  * control}

\item{measurements, nm}{number of measurements}

\item{nscor}{Nonsphericity correction coefficien Sphericity was non-significant = 1}

\item{type}{Type of analysis (0 or 1 or 2). The value "0" is for between-effect; "1" is for within-effect; and "2" is for interaction effect.}

\item{predictors, df, u, v, k, n}{predictors for lm

u    =  degrees of freedom for numerator (pwr.f2.test)
v    = degrees of freedom for denominator (pwr.f2.test)
k    = Number of groups (pwr.anova.test)
n    = Number of observations (per group pwr.anova.test)}

\item{f, f2, r, d}{Effect size. r="Correlation coefficient"
d    = "Standardized difference value (Cohen's d)"
f= "Cohen's f"
f2=    "Cohen's f-squared"}

\item{power}{Statistical power.}

\item{sig.level}{significance level chosed for the test. It equals 0.05 by default.}

\item{alternative}{hypothese}
}
\value{
data.frame
}
\description{
Sample Size
}
\examples{

sample_size("cor", drop.out.rate = .10)
sample_size("aov")
sample_size("lm")
sample_size("aov", treatment = 2 , f = .5)
sample_size("anova", treatment = 2, f = .5)
sample_size("anova", treatment = 2, control = 2 * 2, f = .5)
sample_size("anova", treatment = 1, measurements = 4, f = 0.608)
}
