% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop-out.R
\name{unequal_treatment}
\alias{unequal_treatment}
\title{Unequal Treatment Allocation}
\usage{
unequal_treatment(n, k)
}
\arguments{
\item{n}{is the sample size required}

\item{k}{ratio}
}
\value{
vector
}
\description{
Unequal Treatment Allocation
}
\examples{

unequal_treatment(n= 100, k =2/1)

}
